+++ eagle/entry.go
@@ -87,12 +87,6 @@ func (e *Eagle) PostSaveEntry(ee *entry.Entry) {
 	// Remove entry from the cache. Every other action from here on
 	// should not influence how the entry is rendered.
 	e.RemoveCache(ee)
-
-	// Send webmentions.
-	err = e.SendWebmentions(ee)
-	if err != nil {
-		e.Error(err)
-	}
 }
 
 func (e *Eagle) processPhotos(ee *entry.Entry) error {
